Rename JS files
