+++ challenges.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script type="module" src="src/js/filter_by_type_on_load.js"></script>
+    <script type="module" src="src/js/filterByTypeOnLoad.js"></script>
     <script type="module" src="src/js/tag-filter.js"></script>
     <link rel="stylesheet" href="static/stupidfilter.css">
     <link rel="stylesheet" type="text/css" href="static/main.css" />
