+++ index.html
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css">
     <link rel="stylesheet" type="text/css" href="static/main.css">
-    <script type="module" src="src/js/popular_challenges.js"></script>
+    <script type="module" src="src/js/popularChallenges.js"></script>
     <title>[ESC] Hacker Escape Rooms</title>
 </head>
