+++ src/js/filterByTypeOnLoad.js
