+++ src/js/popularChallenges.js
@@ -1,6 +1,6 @@
 import loadData from "./loadData.js";
 import { createChallengeCard } from "./challenge-card.js";
-import renderChallenges from "./render_challenges.js";
-import filterPopularChallenges from "./filter_popular_challenges.js";
+import renderChallenges from "./renderChallenges.js";
+import filterPopularChallenges from "./filterPopularChallenges.js";
 
 renderChallenges(loadData, filterPopularChallenges, createChallengeCard, document.querySelector(".challenge-list"));
